jsonhkp.PublicKey struct needs to be exported
Because of golang/go#21353, it is no longer possible for
`json.Unmarshal` to set values into unexported structs. This is reported
in golang/go#21357:

`json: cannot set embedded pointer to unexported struct: jsonhkp.publicKey`

The correct solution is to export the struct, because it is a public
interface.

Contributed by @quantcast: https://quantcast.com.